I have defined a Scikit-learn estimator
A Support Vector Machine (SVM) was used to build a binary classification model and detect plagiarization. Scikit-learn SVC is an implementation of a classifier built using an SVM and it is naturally a good fit for classification problems. In this project, due to the smaller size of the dataset (< 10,000 samples), LinearSVC was not needed to optimize training time and the standard SVC implementation was used.